# perf(adapter): temporarily hash large data when formatting chat prompt for speed+memory performance