--- a/dspy/adapters/base.py
+++ b/dspy/adapters/base.py
@@ -8,6 +8,7 @@ import litellm
 from dspy.adapters.types import History, Type
 from dspy.adapters.types.base_type import split_message_content_for_custom_types
 from dspy.adapters.types.tool import Tool, ToolCalls
+from dspy.adapters.utils import LargePayloadHashManager
 from dspy.experimental import Citations
 from dspy.signatures.signature import Signature
 from dspy.utils.callback import BaseCallback, with_callbacks