--- a/dspy/adapters/base.py
+++ b/dspy/adapters/base.py
@@ -1,4 +1,3 @@
-import copy
 import logging
 from typing import TYPE_CHECKING, Any, get_origin