@jmhb0 2025-11-06
The type definitions are in https://github.com/stanfordnlp/dspy/tree/main/dspy/adapters/types 

I don't know about official docs, but most of them seem to have good docstrings.