@tdb-alcorn 2025-12-04
@okhat or @arnavsinghvi11 - can this PR get on track to be merged? dspy is basically unusable with multimedia prompts, this looks like a reasonable direction to fix